Smart Serial: Start to enable parameter setting via HAL
The only way to change settings on the smart-serial remotes is currently
with setsserial. This makes some sense with non-volatile parameters, but
is not useful for (for example) changing the position scale on a servo
drive.
This commit is the first step. At this point all that has changed is the
fact that HAL parameters are now created for any item that appears in the
global list but not in the process data. (previously only nonvol types
were made parameters.


Signed-off-by: andy pugh <andy@bodgesoc.org>